+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## UPDATED FUNCTIONS:
 * `FourPHFfit.bulk` - Fixed false "`total number of seeds tested is less than the total number of germinated seeds`" warning when germination count data is cumulative (Thanks to D. Vijay for vijaydunna[at]gmail.com).
+* `FourPHFfit` - Updated computation of starting values.
 
 ## VIGNETTE:
 * Resolved duplicate chunk labels.